add middleware and reporting finish
