+++ infra/docker-compose.yml
@@ -111,6 +111,13 @@ services:
       NEXT_PUBLIC_TYPESENSE_API_URL: http://localhost:8108
       NEXT_PUBLIC_USE_EMULATOR: "true"
       NEXT_PUBLIC_PROJECT_ID_FOR_TEST: demo-dtp
+      # TODO: remove
+      GCLOUD_PROJECT: "demo-dtp"
+      FIRESTORE_EMULATOR_HOST: "firebase:8080"
+      FIREBASE_AUTH_EMULATOR_HOST: "firebase:9099"
+      FIREBASE_STORAGE_EMULATOR_HOST: "firebase:9199"
+      TYPESENSE_API_URL: "http://firebase:8108"
+      TYPESENSE_API_KEY: "test-api-key"
 
 # volumes:
   # search: {}
